Load spline change and 39 bus introduction
Changed Load_spline to be a single, multi-dimensional structure instead
of an array of structures. Added 39 bus case, as well as a number of
sub-functions to support this inclusion (i.e. replicate_case,
set_ramp_rate, get_ps_areas). However, large oscillations are seen now,
possibly due to the change in definition of B (area frequency bias
variable) within get_ps_areas. When k is made extremely small, this is
fixed, which explains why Will's k was so tiny.